feat: offload graph build to a web worker
Move the heavy graph build (MGF parse, FLASH neighbour search, Louvain and
Leiden clustering, and the ForceAtlas2 layout) off the main thread into a
dedicated wasm web worker so the UI no longer freezes during a rebuild.

Introduce two crates: mascot-web-core holds the shared compute and the serde
message types, and mascot-web-worker is a cdylib that runs the build loop in a
DedicatedWorkerGlobalScope. The app build.rs compiles the worker to wasm,
runs wasm-bindgen over it, and emits the glue plus a generic loader as bundled
assets. worker.rs is the main-thread client that bridges use_resource to the
worker over postMessage with a ready handshake and request ids, and falls back
to building inline if the worker cannot be created.
