+++ crates/mascot-web/.gitignore
@@ -1,0 +1,2 @@
+# Generated by build.rs (compiled worker glue + wasm + loader).
+/assets/generated/
